fix GenerateConnectionFromOpts issue
