+++ internal/models/cli.go
@@ -1,10 +1,14 @@
 package models
 
 import (
+	"fmt"
 	"net/url"
+	"os"
+	"strings"
 	"time"
 
 	"github.com/nats-io/jsm.go/natscontext"
+	"github.com/nats-io/nats.go"
 )
 
 type UiOptions struct {
@@ -150,4 +154,64 @@ type TokenBucket struct {
 	// Required: true
 	// Minimum: 0
 	Size *int64 `json:"size"`
+}
+
+func GenerateConnectionFromOpts(opts *Options) (*nats.Conn, error) {
+	ctxOpts := []natscontext.Option{
+		natscontext.WithServerURL(opts.Servers),
+		natscontext.WithCreds(opts.Creds),
+		natscontext.WithNKey(opts.Nkey),
+		natscontext.WithCertificate(opts.TlsCert),
+		natscontext.WithKey(opts.TlsKey),
+		natscontext.WithCA(opts.TlsCA),
+	}
+
+	if opts.TlsFirst {
+		ctxOpts = append(ctxOpts, natscontext.WithTLSHandshakeFirst())
+	}
+
+	if opts.Username != "" && opts.Password == "" {
+		ctxOpts = append(ctxOpts, natscontext.WithToken(opts.Username))
+	} else {
+		ctxOpts = append(ctxOpts, natscontext.WithUser(opts.Username), natscontext.WithPassword(opts.Password))
+	}
+
+	var err error
+
+	exist, _ := fileAccessible(opts.ConfigurationContext)
+
+	if exist && strings.HasSuffix(opts.ConfigurationContext, ".json") {
+		opts.Configuration, err = natscontext.NewFromFile(opts.ConfigurationContext, ctxOpts...)
+	} else {
+		opts.Configuration, err = natscontext.New(opts.ConfigurationContext, !opts.SkipContexts, ctxOpts...)
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	conn, err := opts.Configuration.Connect()
+	if err != nil {
+		return nil, err
+	}
+	return conn, nil
+}
+
+func fileAccessible(f string) (bool, error) {
+	stat, err := os.Stat(f)
+	if err != nil {
+		return false, err
+	}
+
+	if stat.IsDir() {
+		return false, fmt.Errorf("is a directory")
+	}
+
+	file, err := os.Open(f)
+	if err != nil {
+		return false, err
+	}
+	file.Close()
+
+	return true, nil
 }
