+++ nex/nodes.go
@@ -10,11 +10,12 @@ import (
 
 	"github.com/nats-io/natscli/columns"
 	controlapi "github.com/synadia-io/nex/internal/control-api"
+	"github.com/synadia-io/nex/internal/models"
 )
 
 // Uses a control API client to request a node list from a NATS environment
 func ListNodes(ctx context.Context) error {
-	nc, err := generateConnectionFromOpts()
+	nc, err := models.GenerateConnectionFromOpts(Opts)
 	if err != nil {
 		return err
 	}
@@ -33,8 +34,7 @@ func ListNodes(ctx context.Context) error {
 
 // Uses a control API client to retrieve info on a single node
 func NodeInfo(ctx context.Context, nodeid string) error {
-
-	nc, err := generateConnectionFromOpts()
+	nc, err := models.GenerateConnectionFromOpts(Opts)
 	if err != nil {
 		return err
 	}
